+++ src/Cryptol/Eval/Concrete.hs
@@ -299,7 +299,7 @@ suiteBPrims = Map.fromList $ map (\(n, v) -> (suiteBPrim n, v))
            let f :: Word32 -> Eval Value
                f = pure . VWord . wordVal . BV 32 . toInteger
                zs = finiteSeqMap Concrete (map f [w0,w1,w2,w3,w4,w5,w6])
-           seq zs (pure (VSeq 7 zs)))
+           seq zs (wordSeq Concrete 7 32 zs))
 
   , ("processSHA2_256", {-# SCC "SuiteB::processSHA2_256" #-}
      PFinPoly \n ->
@@ -312,7 +312,7 @@ suiteBPrims = Map.fromList $ map (\(n, v) -> (suiteBPrim n, v))
            let f :: Word32 -> Eval Value
                f = pure . VWord . wordVal . BV 32 . toInteger
                zs = finiteSeqMap Concrete (map f [w0,w1,w2,w3,w4,w5,w6,w7])
-           seq zs (pure (VSeq 8 zs)))
+           seq zs (wordSeq Concrete 8 32 zs))
 
   , ("processSHA2_384", {-# SCC "SuiteB::processSHA2_384" #-}
      PFinPoly \n ->
@@ -325,7 +325,7 @@ suiteBPrims = Map.fromList $ map (\(n, v) -> (suiteBPrim n, v))
            let f :: Word64 -> Eval Value
                f = pure . VWord . wordVal . BV 64 . toInteger
                zs = finiteSeqMap Concrete (map f [w0,w1,w2,w3,w4,w5])
-           seq zs (pure (VSeq 6 zs)))
+           seq zs (wordSeq Concrete 6 64 zs))
 
   , ("processSHA2_512", {-# SCC "SuiteB::processSHA2_512" #-}
      PFinPoly \n ->
@@ -338,7 +338,7 @@ suiteBPrims = Map.fromList $ map (\(n, v) -> (suiteBPrim n, v))
            let f :: Word64 -> Eval Value
                f = pure . VWord . wordVal . BV 64 . toInteger
                zs = finiteSeqMap Concrete (map f [w0,w1,w2,w3,w4,w5,w6,w7])
-           seq zs (pure (VSeq 8 zs)))
+           seq zs (wordSeq Concrete 8 64 zs))
 
   , ("AESKeyExpand", {-# SCC "SuiteB::AESKeyExpand" #-}
       PFinPoly \k ->
@@ -352,7 +352,7 @@ suiteBPrims = Map.fromList $ map (\(n, v) -> (suiteBPrim n, v))
             kws <- mapM toWord [0 .. k-1]
             let ws = AES.keyExpansionWords k kws
             let len = 4*(k+7)
-            pure (VSeq len (finiteSeqMap Concrete (map fromWord ws))))
+            wordSeq Concrete len 32 (finiteSeqMap Concrete (map fromWord ws)))
 
   , ("AESInvMixColumns", {-# SCC "SuiteB::AESInvMixColumns" #-}
       PFun \st ->
@@ -364,7 +364,7 @@ suiteBPrims = Map.fromList $ map (\(n, v) -> (suiteBPrim n, v))
                 fromWord = pure . VWord . wordVal . BV 32 . toInteger
             ws <- mapM toWord [0,1,2,3]
             let ws' = AES.invMixColumns ws
-            pure . VSeq 4 . finiteSeqMap Concrete . map fromWord $ ws')
+            wordSeq Concrete 4 32 (finiteSeqMap Concrete . map fromWord $ ws'))
 
   , ("AESEncRound", {-# SCC "SuiteB::AESEncRound" #-}
       PFun \st ->
@@ -376,7 +376,7 @@ suiteBPrims = Map.fromList $ map (\(n, v) -> (suiteBPrim n, v))
                 fromWord = pure . VWord . wordVal . BV 32 . toInteger
             ws <- mapM toWord [0,1,2,3]
             let ws' = AES.aesRound ws
-            pure . VSeq 4 . finiteSeqMap Concrete . map fromWord $ ws')
+            wordSeq Concrete 4 32 (finiteSeqMap Concrete . map fromWord $ ws'))
 
   , ("AESEncFinalRound", {-# SCC "SuiteB::AESEncFinalRound" #-}
      PFun \st ->
@@ -388,7 +388,7 @@ suiteBPrims = Map.fromList $ map (\(n, v) -> (suiteBPrim n, v))
                 fromWord = pure . VWord . wordVal . BV 32 . toInteger
             ws <- mapM toWord [0,1,2,3]
             let ws' = AES.aesFinalRound ws
-            pure . VSeq 4 . finiteSeqMap Concrete . map fromWord $ ws')
+            wordSeq Concrete 4 32 (finiteSeqMap Concrete . map fromWord $ ws'))
 
   , ("AESDecRound", {-# SCC "SuiteB::AESDecRound" #-}
       PFun \st ->
@@ -400,7 +400,7 @@ suiteBPrims = Map.fromList $ map (\(n, v) -> (suiteBPrim n, v))
                 fromWord = pure . VWord . wordVal . BV 32 . toInteger
             ws <- mapM toWord [0,1,2,3]
             let ws' = AES.aesInvRound ws
-            pure . VSeq 4 . finiteSeqMap Concrete . map fromWord $ ws')
+            wordSeq Concrete 4 32 . finiteSeqMap Concrete . map fromWord $ ws')
 
   , ("AESDecFinalRound", {-# SCC "SuiteB::AESDecFinalRound" #-}
      PFun \st ->
@@ -412,7 +412,7 @@ suiteBPrims = Map.fromList $ map (\(n, v) -> (suiteBPrim n, v))
                 fromWord = pure . VWord . wordVal . BV 32 . toInteger
             ws <- mapM toWord [0,1,2,3]
             let ws' = AES.aesInvFinalRound ws
-            pure . VSeq 4 . finiteSeqMap Concrete . map fromWord $ ws')
+            wordSeq Concrete 4 32 . finiteSeqMap Concrete . map fromWord $ ws')
   ]