+++ src/Cryptol/Symbolic.hs
@@ -56,7 +56,6 @@ import qualified LibBF as FP
 
 import           Cryptol.Backend
 import           Cryptol.Backend.FloatHelpers(bfValue)
-import           Cryptol.Backend.SeqMap (finiteSeqMap)
 import           Cryptol.Backend.WordValue (wordVal)
 
 import qualified Cryptol.Eval.Concrete as Concrete
@@ -270,7 +269,9 @@ varShapeToValue sym var =
     VarRational n d -> VRational (SRational n d)
     VarWord w    -> VWord (wordVal w)
     VarFloat f   -> VFloat f
-    VarFinSeq n vs -> VSeq n (finiteSeqMap sym (map (pure . varShapeToValue sym) vs))
+    VarFinSeq n vs -> case toFinSeq (map (varShapeToValue sym) vs) of
+      Just vs' -> finSeq sym n vs'
+      Nothing -> panic "varShapeToValue" ["Unexpected VarBit in VarFinSeq"]
     VarTuple vs  -> VTuple (map (pure . varShapeToValue sym) vs)
     VarRecord fs -> VRecord (fmap (pure . varShapeToValue sym) fs)
     VarEnum tag cons ->